src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
weasyprint==62.3
python-multipart==0.0.9
